refactor(ui): merge SSH key utils into single camelCase file
Two files clashed: ssh-keys.ts (validation, fingerprinting, signing) and
sshKeys.ts (isPublicKeyValid), making it ambiguous which to import. Merge
both into sshKeys.ts to match the project's camelCase convention. No logic
changes.

Fixes: shellhub-io/team#141
